Rename dg_uty.c
