+++ main.lsf
@@ -191,7 +191,7 @@ Overlay OVY_29
 	Object asm/overlay_29_022E330C.o
 	Object src/dg_effect_022E37B8.o
 	Object asm/overlay_29_022E37DC.o
-	Object src/overlay_29_022E95F4.o
+	Object src/dg_uty_022E95F4.o
 	Object asm/overlay_29_022E9618.o
 	Object src/dg_random_022EAB20.o
 	Object asm/overlay_29_022EAB50.o
